third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-getFlattenedDocument-expected.txt

A
B
C
D
E

{
  "nodes": [
    {
      "nodeId": 4,
      "parentId": 3,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "SCRIPT",
      "localName": "script",
      "nodeValue": "",
      "childNodeCount": 0,
      "children": [],
      "attributes": [
        "type",
        "text/javascript",
        "src",
        "../../http/tests/inspector-protocol/inspector-protocol-test.js"
      ]
    },
    {
      "nodeId": 6,
      "parentId": 5,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": "\n\nfunction test()\n{\n    InspectorTest.sendCommand(\"DOM.enable\", {});\n    InspectorTest.sendCommandOrDie(\"DOM.getFlattenedDocument\", {\"depth\": -1, \"pierce\": true}, onDocument);\n\n    function onDocument(response) {\n        function stabilize(key, value) {\n            var unstableKeys = [\"backendNodeId\", \"documentURL\", \"baseURL\", \"frameId\"];\n            if (unstableKeys.indexOf(key) !== -1)\n                return \"<\" + typeof(value) + \">\";\n            return value;\n        }\n        InspectorTest.log(JSON.stringify(response, stabilize, 2));\n        InspectorTest.completeTest();\n    }\n}\n\n"
    },
    {
      "nodeId": 5,
      "parentId": 3,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "SCRIPT",
      "localName": "script",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 3,
      "parentId": 2,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "HEAD",
      "localName": "head",
      "nodeValue": "",
      "childNodeCount": 2,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 9,
      "parentId": 8,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": " A\n      "
    },
    {
      "nodeId": 11,
      "parentId": 10,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": " B\n        "
    },
    {
      "nodeId": 13,
      "parentId": 12,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": " C\n          "
    },
    {
      "nodeId": 15,
      "parentId": 14,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": " D\n            "
    },
    {
      "nodeId": 17,
      "parentId": 16,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": " E\n            "
    },
    {
      "nodeId": 16,
      "parentId": 14,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": [
        "id",
        "E"
      ]
    },
    {
      "nodeId": 14,
      "parentId": 12,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 2,
      "children": [],
      "attributes": [
        "id",
        "D"
      ]
    },
    {
      "nodeId": 12,
      "parentId": 10,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 2,
      "children": [],
      "attributes": [
        "id",
        "C"
      ]
    },
    {
      "nodeId": 10,
      "parentId": 8,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 2,
      "children": [],
      "attributes": [
        "id",
        "B"
      ]
    },
    {
      "nodeId": 8,
      "parentId": 7,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 2,
      "children": [],
      "attributes": [
        "id",
        "A"
      ]
    },
    {
      "nodeId": 23,
      "parentId": 22,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": "\n@font-face {\n  font-family: 'ahem';\n  src: url(../../../resources/Ahem.ttf);\n}\n"
    },
    {
      "nodeId": 22,
      "parentId": 21,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "STYLE",
      "localName": "style",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 21,
      "parentId": 20,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "HEAD",
      "localName": "head",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 26,
      "parentId": 25,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": "\nHello from the iframe.\n"
    },
    {
      "nodeId": 25,
      "parentId": 24,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": [
        "style",
        "font-family: ahem;"
      ]
    },
    {
      "nodeId": 24,
      "parentId": 20,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "BODY",
      "localName": "body",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 20,
      "parentId": 19,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "HTML",
      "localName": "html",
      "nodeValue": "",
      "childNodeCount": 2,
      "children": [],
      "attributes": [],
      "frameId": "<string>"
    },
    {
      "nodeId": 18,
      "parentId": 7,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "IFRAME",
      "localName": "iframe",
      "nodeValue": "",
      "childNodeCount": 0,
      "children": [],
      "attributes": [
        "src",
        "../dom/resources/simple-iframe.html",
        "width",
        "400",
        "height",
        "200"
      ],
      "frameId": "<string>",
      "contentDocument": {
        "nodeId": 19,
        "backendNodeId": "<number>",
        "nodeType": 9,
        "nodeName": "#document",
        "localName": "",
        "nodeValue": "",
        "childNodeCount": 1,
        "children": [],
        "documentURL": "<string>",
        "baseURL": "<string>",
        "xmlVersion": ""
      }
    },
    {
      "nodeId": 30,
      "parentId": 29,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": "\n:host {\n    color: red;\n}\n"
    },
    {
      "nodeId": 29,
      "parentId": 28,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "STYLE",
      "localName": "style",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 31,
      "parentId": 28,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 0,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 33,
      "parentId": 32,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": "Hi from a template!"
    },
    {
      "nodeId": 32,
      "parentId": 28,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "H1",
      "localName": "h1",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": []
    },
    {
      "nodeId": 27,
      "parentId": 7,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "DIV",
      "localName": "div",
      "nodeValue": "",
      "childNodeCount": 0,
      "children": [],
      "attributes": [
        "id",
        "shadow-host"
      ],
      "shadowRoots": [
        {
          "nodeId": 28,
          "backendNodeId": "<number>",
          "nodeType": 11,
          "nodeName": "#document-fragment",
          "localName": "",
          "nodeValue": "",
          "childNodeCount": 3,
          "children": [],
          "shadowRootType": "open"
        }
      ]
    },
    {
      "nodeId": 35,
      "parentId": 34,
      "backendNodeId": "<number>",
      "nodeType": 3,
      "nodeName": "#text",
      "localName": "",
      "nodeValue": "\n        var host = document.querySelector(\"#shadow-host\").createShadowRoot();\n        var template = document.querySelector(\"#shadow-template\");\n        host.appendChild(template.content);\n        template.remove();\n        window.onload = runTest;\n    "
    },
    {
      "nodeId": 34,
      "parentId": 7,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "SCRIPT",
      "localName": "script",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "attributes": [
        "type",
        "text/javascript"
      ]
    },
    {
      "nodeId": 7,
      "parentId": 2,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "BODY",
      "localName": "body",
      "nodeValue": "",
      "childNodeCount": 4,
      "children": [],
      "attributes": [
        "class",
        "body-class"
      ]
    },
    {
      "nodeId": 2,
      "parentId": 1,
      "backendNodeId": "<number>",
      "nodeType": 1,
      "nodeName": "HTML",
      "localName": "html",
      "nodeValue": "",
      "childNodeCount": 2,
      "children": [],
      "attributes": [],
      "frameId": "<string>"
    },
    {
      "nodeId": 1,
      "backendNodeId": "<number>",
      "nodeType": 9,
      "nodeName": "#document",
      "localName": "",
      "nodeValue": "",
      "childNodeCount": 1,
      "children": [],
      "documentURL": "<string>",
      "baseURL": "<string>",
      "xmlVersion": ""
    }
  ]
}
